requirements.txt
python-dotenv==1.0.0
pydantic==2.6.4
fastapi==0.115.0
uvicorn[standard]==0.30.0
openai==0.27.14
requests==2.32.3
httpx==0.27.0
